lexer with comments

+++ srcs/lexer_utils.c
@@ -1,12 +1,30 @@
 #include "lexer.h"
 
-size_t	is_eoi(char c, t_token *token)
+/* Fonction qui detect le token End Of Input (\0)
+ * @param: c
+ * - le caracter a evaluer
+ *
+ * @param: token
+ * - Pointeur vers le token a modifier
+ *
+ * @return: int
+ * - token type
+ * */
+int	is_eoi(char c, t_token *token)
 {
 	if (!c)
 		token->type = E_EOI;
 	return (token->type);
 }
 
+/* Fonction pour detecter un char qui pourrait
+ * etre le debut d'un token autre que E_WORD
+ * @param: c
+ * - Le caracter a evaluer
+ *
+ * @return: char
+ * - Le char detecte, 0 sinon
+ * */
 int	is_special_token(char c)
 {
 	if (c == '$'
@@ -21,6 +39,16 @@ int	is_special_token(char c)
 	return (0);
 }
 
+/* Fonction pour generer les tokens E_WORD
+ * @param: input
+ * - La chaine a tokeniser
+ *
+ * @param: token
+ * - L'adresse du token a utiliser
+ *
+ * @return: void
+ * - Pas de valeur de retour necessaire pour le moment (a voir pour la gestion de erreurs)
+ * */
 void	word_token(char *input, t_token *token)
 {
 	char	*tmp;
@@ -32,6 +60,17 @@ void	word_token(char *input, t_token *token)
 	token->type  = E_WORD;
 }
 
+/* Fonction pour gerer les tokens E_SEP
+ * @param: input
+ * - La chaine a tokeniser
+ *
+ * @param: token
+ * - L'adresse du token a utiliser
+ *
+ * @return: void
+ * - Pas de valeur de retour necessaire pour le moment (a voir pour la gestion de erreurs)
+ * */
+/*
 void	sep_token(char *input, t_token *token)
 {
 	char *tmp;
@@ -42,8 +81,20 @@ void	sep_token(char *input, t_token *token)
 	token->len = input - tmp;
 	token->type = E_SEP;
 }
+*/
 
-size_t	is_token_1(char *input, t_token *token)
+
+/* Deux fonction pour differencier les tokens != E_WORD
+ * @param: input
+ * - La chaine a tokeniser
+ *
+ * @param: token
+ * - L'adresse du token a utiliser
+ *
+ * @return: int (unsigned int ?)
+ * - Type du token
+ * */
+int	is_token_1(char *input, t_token *token)
 {
 	if (*input == '>')
 		token->type  = E_GREAT;
@@ -60,7 +111,7 @@ size_t	is_token_1(char *input, t_token *token)
 	return (token->type);
 }
 
-size_t	is_token_2(char *input, t_token *token)
+int	is_token_2(char *input, t_token *token)
 {
 	if (!ft_strncmp("&&", input, 2))
 		token->type  = E_AND_IF;
